fix: move ViewLayers and Drawables to core/
