+++ src/app/GUI/mainwindow.h
@@ -48,10 +48,10 @@
 #include "fileshandler.h"
 #include "ekeyfilter.h"
 #include "GUI/RenderWidgets/renderwidget.h"
-#include "viewlayers/viewlayer_preview.h"
-#include "viewlayers/viewlayer_render.h"
-#include "viewlayers/viewlayer_selection.h"
-#include "viewlayers/basecanvas.h"
+#include "ViewLayers/viewlayer_preview.h"
+#include "ViewLayers/viewlayer_render.h"
+#include "ViewLayers/viewlayer_selection.h"
+#include "ViewLayers/basecanvas.h"
 #include "widgets/colortoolbar.h"
 #include "widgets/qdoubleslider.h"
 #include "widgets/canvastoolbar.h"
