+++ src/core/Drawables/rectangle_drawable.h
