+++ src/core/ViewLayers/basecanvas.cpp
