+++ src/core/ViewLayers/utils.cpp
