+++ src/core/ViewLayers/viewlayer_preview.cpp
@@ -32,10 +32,10 @@
 
 #include "basecanvas.h"
 #include "utils.h"
-#include "../drawables/rectangle_drawable.h"
-#include "../../core/Boxes/containerbox.h"
-#include "../../core/efiltersettings.h"
-#include "../../core/GUI/global.h"
+#include "../Drawables/rectangle_drawable.h"
+#include "../Boxes/containerbox.h"
+#include "../efiltersettings.h"
+#include "../GUI/global.h"
 
 
 ViewLayerPreview::ViewLayerPreview(Document &document, BaseCanvas *canvas)
