+++ src/core/ViewLayers/viewlayer_preview.h
