+++ src/ui/CMakeLists.txt
@@ -64,11 +64,6 @@ set(
     optimalscrollarena/scrollwidget.cpp
     optimalscrollarena/scrollwidgetvisiblepart.cpp
     optimalscrollarena/singlewidget.cpp
-    viewlayers/basecanvas.cpp
-    viewlayers/utils.cpp
-    viewlayers/viewlayer_render.cpp
-    viewlayers/viewlayer_preview.cpp
-    viewlayers/viewlayer_selection.cpp
     widgets/aboutwidget.cpp
     widgets/actionbutton.cpp
     widgets/alignwidget.cpp
@@ -129,8 +124,6 @@ set(
     dialogs/qrealpointvaluedialog.h
     dialogs/renderoutputwidget.h
     dialogs/scenesettingsdialog.h
-    drawables/drawable.h
-    drawables/rectangle_drawable.h
     gradientwidgets/currentgradientwidget.h
     gradientwidgets/displayedgradientswidget.h
     gradientwidgets/gradientslistwidget.h
@@ -143,12 +136,6 @@ set(
     optimalscrollarena/scrollwidget.h
     optimalscrollarena/scrollwidgetvisiblepart.h
     optimalscrollarena/singlewidget.h
-    viewlayers/utils.h
-    viewlayers/basecanvas.h
-    viewlayers/viewlayer.h
-    viewlayers/viewlayer_render.h
-    viewlayers/viewlayer_preview.h
-    viewlayers/viewlayer_selection.h
     widgets/aboutwidget.h
     widgets/actionbutton.h
     widgets/alignwidget.h
